# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: PaperFull API
  description: |
    AI-powered academic paper generation API. Authentication via httpOnly cookies
    + double-submit CSRF token (X-CSRF-TOKEN header for state-changing requests).
  version: "1.1.0"
  contact:
    name: PaperFull
    url: https://paperfull.app
servers:
  - url: https://paperfull.app
    description: Production
  - url: http://localhost:8000
    description: Local proxy (frontend)
  - url: http://localhost:8001
    description: Local backend (direct)
...
paths:
...
  /api/images/{paper_id}/{filename}:
    parameters:
      - $ref: "#/components/parameters/PaperId"
      - name: filename
        in: path
        required: true
        schema: { type: string, pattern: "^[A-Za-z0-9_.-]{1,128}$" }
      - name: s
        in: query
        schema: { type: string }
      - name: t
        in: query
        schema: { type: string }
    get:
      tags: [Images]
      summary: Serve image bytes
      operationId: getImage
      security:
        - {}
        - cookieAuth: []
      responses:
        "200":
          description: Image file bytes
          content:
            image/png:
              schema: { type: string, format: binary }
            image/jpeg:
              schema: { type: string, format: binary }
            image/gif:
              schema: { type: string, format: binary }
            image/webp:
              schema: { type: string, format: binary }
        "404":
          description: Image not found
...
components:
  parameters:
    PaperId:
      name: paper_id
      in: path
      required: true
      schema: { type: string, pattern: "^[A-Za-z0-9_-]{1,64}$" }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: |
        httpOnly access cookie. CSRF protection: include the value of the
        `csrf_access_token` cookie as the `X-CSRF-TOKEN` header on all
        state-changing requests (POST/PUT/PATCH/DELETE).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Legacy header-based auth (still accepted as fallback).
security:
  - cookieAuth: []
  - bearerAuth: []